requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pandas>=2.0.0
numpy>=2.0.0
scikit-learn>=1.4.0
joblib>=1.3.0
python-jose[cryptography]>=3.3.0
slowapi>=0.1.9
prometheus-fastapi-instrumentator>=7.0.0
python-docx>=1.1.0

